+++ cmd/src/lsif_upload.go
@@ -4,14 +4,15 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
+	"net/http"
 	"net/url"
 	"os"
+	"sort"
 	"strings"
 	"sync"
 	"time"
 
 	"github.com/efritz/pentimento"
-	"github.com/mattn/go-isatty"
 	"github.com/pkg/browser"
 	"github.com/pkg/errors"
 	"github.com/sourcegraph/codeintelutils"
@@ -52,6 +53,9 @@ Examples:
 		maxPayloadSizeMb     *int
 		ignoreUploadFailures *bool
 		uploadRoute          *string
+		rawVerbosity         *int
+		verbosity            lsifUploadVerbosity
+		associatedIndexID    *int
 	}
 
 	flagSet := flag.NewFlagSet("upload", flag.ExitOnError)
@@ -66,10 +70,14 @@ Examples:
 	flags.noProgress = flagSet.Bool("no-progress", false, `Do not display a progress bar.`)
 	flags.maxPayloadSizeMb = flagSet.Int("max-payload-size", 100, `The maximum upload size (in megabytes). Indexes exceeding this limit will be uploaded over multiple HTTP requests.`)
 	flags.ignoreUploadFailures = flagSet.Bool("ignore-upload-failure", false, `Exit with status code zero on upload failure.`)
-	flags.uploadRoute = flagSet.String("upload-route", "/.api/lsif/upload", "The path of the upload route.")
+	flags.uploadRoute = flagSet.String("upload-route", "/.api/lsif/upload", "The path of the upload route. For internal use only.")
+	flags.rawVerbosity = flagSet.Int("trace", 0, "-trace=0 shows no logs; -trace=1 shows requests and response metadata; -trace=2 shows headers, -trace=3 shows response body")
+	flags.associatedIndexID = flagSet.Int("associated-index-id", -1, "ID of the associated index record for this upload. For internal use only.")
 
 	parseAndValidateFlags := func(args []string) error {
-		flagSet.Parse(args)
+		if err := flagSet.Parse(args); err != nil {
+			return err
+		}
 
 		type inferError struct {
 			argument string
@@ -146,8 +154,18 @@ Examples:
 			return errors.New("max-payload-size must be positive")
 		}
 
+		// Don't need to check upper bounds as we only compare verbosity ranges
+		// It's fine if someone supplies -trace=42, but it will just behave the
+		// same as if they supplied the highest verbosity level we define
+		// internally.
+		flags.verbosity = lsifUploadVerbosity(*flags.rawVerbosity)
+
 		if !*flags.json {
 			fmt.Println(argsString)
+		}
+
+		if *flags.associatedIndexID < 0 {
+			flags.associatedIndexID = nil
 		}
 
 		return nil
@@ -170,9 +188,11 @@ Examples:
 			GitHubToken:          *flags.gitHubToken,
 			File:                 *flags.file,
 			MaxPayloadSizeBytes:  *flags.maxPayloadSizeMb * 1000 * 1000,
+			AssociatedIndexID:    flags.associatedIndexID,
 			MaxRetries:           10,
 			RetryInterval:        time.Millisecond * 250,
 			UploadProgressEvents: make(chan codeintelutils.UploadProgressEvent),
+			Logger:               &lsifUploadRequestLogger{verbosity: flags.verbosity},
 		}
 
 		var wg sync.WaitGroup
@@ -181,15 +201,15 @@ Examples:
 		go func() {
 			defer wg.Done()
 
-			if *flags.json || *flags.noProgress {
+			if *flags.json || *flags.noProgress || flags.verbosity > 0 {
 				return
 			}
 
-			pentimento.PrintProgress(func(p *pentimento.Printer) error {
+			_ = pentimento.PrintProgress(func(p *pentimento.Printer) error {
 				for event := range opts.UploadProgressEvents {
 					content := pentimento.NewContent()
 					content.AddLine(formatProgressBar(event.TotalProgress, fmt.Sprintf("%d/%d", event.Part, event.NumParts)))
-					p.WriteContent(content)
+					_ = p.WriteContent(content)
 				}
 
 				_ = p.Reset()
@@ -202,19 +222,19 @@ Examples:
 		wg.Wait()                        // Wait for progress bar goroutine to clear screen
 		if err != nil {
 			if err == codeintelutils.ErrUnauthorized {
-				if *flags.gitHubToken == "" {
-					return fmt.Errorf("you must provide -github-token=TOKEN, where TOKEN is a GitHub personal access token with 'repo' or 'public_repo' scope")
+				err = errorWithHint{
+					err: err, hint: strings.Join([]string{
+						"You may need to specify or update your GitHub access token to use this endpoint.",
+						"See https://docs.sourcegraph.com/cli/references/lsif/upload.",
+					}, "\n"),
 				}
 
-				if isatty.IsTerminal(os.Stdout.Fd()) {
-					fmt.Println("You may need to specify or update your GitHub access token to use this endpoint.")
-					fmt.Println("See https://github.com/sourcegraph/src-cli#authentication.")
-				}
 			}
 
 			if *flags.ignoreUploadFailures {
-				fmt.Printf("error: %s\n", err)
-				return nil
+				// Report but don't return
+				fmt.Println(err.Error())
+				err = nil
 			}
 
 			return err
@@ -305,10 +325,75 @@ func formatProgressBar(progress float64, suffix string) string {
 	)
 }
 
-// digits returns the number of digits of n.
-func digits(n int) int {
-	if n >= 10 {
-		return 1 + digits(n/10)
+type errorWithHint struct {
+	err  error
+	hint string
+}
+
+func (e errorWithHint) Error() string {
+	return fmt.Sprintf("error: %s\n\n%s\n", e.err, e.hint)
+}
+
+type lsifUploadVerbosity int
+
+const (
+	lsifUploadVerbosityNone                  lsifUploadVerbosity = iota // -trace=0 (default)
+	lsifUploadVerbosityTrace                                            // -trace=1
+	lsifUploadVerbosityTraceShowHeaders                                 // -trace=2
+	lsifUploadVerbosityTraceShowResponseBody                            // -trace=3
+)
+
+type lsifUploadRequestLogger struct {
+	verbosity lsifUploadVerbosity
+}
+
+func (l *lsifUploadRequestLogger) LogRequest(req *http.Request) {
+	if l.verbosity == lsifUploadVerbosityNone {
+		return
 	}
-	return 1
+
+	if l.verbosity >= lsifUploadVerbosityTrace {
+		fmt.Printf("> %s %s\n", req.Method, req.URL)
+	}
+
+	if l.verbosity >= lsifUploadVerbosityTraceShowHeaders {
+		fmt.Printf("> Request Headers:\n")
+		for _, k := range sortHeaders(req.Header) {
+			fmt.Printf(">     %s: %s\n", k, req.Header[k])
+		}
+	}
+
+	fmt.Printf("\n")
+}
+
+func (l *lsifUploadRequestLogger) LogResponse(req *http.Request, resp *http.Response, body []byte, elapsed time.Duration) {
+	if l.verbosity == lsifUploadVerbosityNone {
+		return
+	}
+
+	if l.verbosity >= lsifUploadVerbosityTrace {
+		fmt.Printf("< %s %s %s in %s\n", req.Method, req.URL, resp.Status, elapsed)
+	}
+
+	if l.verbosity >= lsifUploadVerbosityTraceShowHeaders {
+		fmt.Printf("< Response Headers:\n")
+		for _, k := range sortHeaders(resp.Header) {
+			fmt.Printf("<     %s: %s\n", k, resp.Header[k])
+		}
+	}
+
+	if l.verbosity >= lsifUploadVerbosityTraceShowResponseBody {
+		fmt.Printf("< Response Body: %s\n", body)
+	}
+
+	fmt.Printf("\n")
+}
+
+func sortHeaders(header http.Header) []string {
+	var keys []string
+	for k := range header {
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+	return keys
 }